+++ src/main/java/com/lightstreamer/adapters/remote/Server.java
@@ -172,16 +172,18 @@ public abstract class Server {
     }
 
     /** 
-     * Sets the stream used by the Remote Data Adapter in order to send asynchronous
-     * data to the Proxy Adapter.
-     * Depending on the Proxy Data Adapter configuration, this stream should be
-     * either the same object provided to setReplyStream or a different stream.
-     * In the first case, the outbound messages pertaining to the "Reply" and
-     * "Notify" streams will be merged properly.
-     * Currently not used and not needed by the Remote Metadata Adapter.
-     * 
-     * @param outputStream the stream used by the Remote Adapter in order to send asynchronous
-     * data to the Proxy Adapter. 
+     * To be used only when connecting with an old Proxy Data Adapter,
+     * based on two-connections behavior; this allows for backward
+     * compatibility with Server version earlier than 7.4.
+     * Sets the stream that should be used by the Remote Data Adapter
+     * in order to send asynchronous data to the Proxy Data Adapter.
+     * In fact, the configuration for old Proxy Data Adapters required the
+     * specification of a "notify" port, to which a second connection
+     * should be opened, with its own stream.<BR>
+     * Not used and not needed by the Remote Metadata Adapter.
+     *  
+     * @param outputStream the stream to be used by the Remote Adapter
+     * in order to send asynchronous data to an old Proxy Data Adapter. 
      */
     public final void setNotifyStream(@Nonnull OutputStream outputStream) {
         if (startedOnce) {
@@ -191,11 +193,11 @@ public abstract class Server {
     }
     /** 
      * Gets the stream used by the Remote Adapter in order to send asynchronous
-     * data to the Proxy Adapter. Currently not used and not needed
-     * by the Remote Metadata Adapter.
+     * data to an old Proxy Data Adapter.<BR>
+     * Not used and not needed by the Remote Metadata Adapter.
      * 
-     * @return the stream used by the Remote Adapter in order to send asynchronous
-     * data to the Proxy Adapter. 
+     * @return the stream used by the Remote Adapter in order to send
+     * asynchronous data to an old Proxy Data Adapter, or null. 
      */
     @Nonnull
     public final OutputStream getNotifyStream() {